+++ observers/state_estimation_cascade/include/state_estimation_cascade/state_estimator.hpp
@@ -91,6 +91,6 @@ private:
   void gpsPositionCb(const GpsMsg::ConstPtr& gps);
 
   void checkTopicsTimerCb(const ros::TimerEvent&);
-  void dynamicReconfigureCb(const ConfigType& cfg, uint32_t);
+  void dynamicReconfigureCb(const ConfigType& cfg, size_t);
 };
 }  // namespace state_estimation_cascade
