+++ observers/state_estimation_eskf/include/state_estimation_eskf/eskf_ros.hpp
@@ -125,6 +125,6 @@ private:
   void gpsCb(const GpsMsg::ConstPtr& gps);
 
   void checkTopicsTimerCb(const ros::TimerEvent&);
-  void dynamicReconfigureCb(const ConfigType& cfg, uint32_t);
+  void dynamicReconfigureCb(const ConfigType& cfg, size_t);
 };
 }  // namespace state_estimation_eskf
