+++ observers/state_estimation_eskf/src/eskf_ros.cpp
@@ -487,7 +487,7 @@ void ErrorStateKalmanFilterRos::checkTopicsTimerCb(const ros::TimerEvent&)
     rosWarn(name_, nh_.getNamespace() << "/" << tobas::kGpsTopic << " is not received yet.");
 }
 
-void ErrorStateKalmanFilterRos::dynamicReconfigureCb(const ConfigType& cfg, uint32_t)
+void ErrorStateKalmanFilterRos::dynamicReconfigureCb(const ConfigType& cfg, size_t)
 {
   grav_cov_.diagonal().fill(cfg.gravity_variance);
   yaw_var_ = cfg.yaw_variance;
