+++ tobas_mr_wind_estimation/include/tobas_mr_wind_estimation/wind_estimator.hpp
@@ -15,7 +15,7 @@ namespace tobas_mr_wind_estimation
 {
 class WindEstimator : public tobas::BaseNode
 {
-  static constexpr uint32_t kStateSize = 2;
+  static constexpr size_t kStateSize = 2;
   static constexpr double kInitWindStddev = 10.;  // [m/s]
 
   using self = WindEstimator;
